requirements.txt
flask==1.0.3
python-bidi==0.4.0
arabic-reshaper==2.0.14
jdatetime==3.2.0
pandas==0.24.2
matplotlib
image==1.5.27
client==0.0.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.14.1
itsdangerous==0.24
suds-py3==1.3.4.0
requests
#wsgiref==0.1.2